git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/mlxxxidl/mlcuddidl/branches/release-3.0@299 c27d38ec-fa33-0410-b621-fb01b97e5d69

+++ cuddaux/cuddauxMisc.c
@@ -52,7 +52,7 @@
 static int cuddaux_list_add(cuddaux_list_t** const plist, DdNode* node);
 static int cuddauxNodesBelowLevelRecur(DdManager* manager, DdNode* F, int level,
 				       cuddaux_list_t** plist, st_table* visited,
-				       size_t max, int *psize,
+				       int max, int *psize,
 				       bool take_background);
 
 
@@ -499,7 +499,7 @@ cuddauxAddGuardOfNodeRecur(DdManager* manager, DdNode* f, DdNode* h)
 static int
 cuddauxNodesBelowLevelRecur(DdManager* manager, DdNode* F, int level,
 			    cuddaux_list_t** plist, st_table* visited,
-			    size_t max, size_t* psize,
+			    int max, int* psize,
 			    bool take_background)
 {
   int topF,res;
